+++ feature/chat/src/main/kotlin/uk/gov/govuk/chat/ui/component/ChatEntry.kt
@@ -1,4 +1,4 @@
-package uk.gov.govuk.chat.ui.chat
+package uk.gov.govuk.chat.ui.component
 
 import android.graphics.ImageDecoder
 import android.graphics.drawable.AnimatedImageDrawable
