+++ feature/chat/src/main/kotlin/uk/gov/govuk/chat/ui/component/ErrorScreens.kt
@@ -1,4 +1,4 @@
-package uk.gov.govuk.chat.ui.chat
+package uk.gov.govuk.chat.ui.component
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.IntrinsicSize
