+++ feature/chat/src/main/kotlin/uk/gov/govuk/chat/ui/component/Markdown.kt
@@ -1,4 +1,4 @@
-package uk.gov.govuk.chat.ui.chat
+package uk.gov.govuk.chat.ui.component
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
